put logging methods in own file, use in more places
