+++ src/Storage.cpp
@@ -6,12 +6,12 @@
 
 #include <Arduino.h>
 #include <EEPROM.h>
+
+#include "debugSerial.h"
 #include "userConfig.h"
 #include "Storage.h"
 
-
 #define STRUCT_MEMBER_SIZE(Type, Member) sizeof(((Type*)0)->Member)
-
 
 // storage data structure
 typedef struct __attribute__((packed)) {
@@ -239,7 +239,7 @@ static inline int32_t getItemAddr(sto_item_id_t itemId, uint16_t* maxItemSize = 
             break;
 
         default:
-            Serial.printf("%s(): invalid item ID %i!\n", __FUNCTION__, itemId);
+            debugPrintf("%s(): invalid item ID %i!\n", __FUNCTION__, itemId);
             addr = -1;
             size = 0;
             break;
@@ -297,7 +297,7 @@ static inline bool isString(const void* buf, uint16_t bufSize) {
  * @brief Sets the default values.
  */
 static void setDefaults(void) {
-  Serial.printf("%s(): %p <- %p (%u)\n", __FUNCTION__, EEPROM.getDataPtr(), &itemDefaults, sizeof(itemDefaults));
+  debugPrintf("%s(): %p <- %p (%u)\n", __FUNCTION__, EEPROM.getDataPtr(), &itemDefaults, sizeof(itemDefaults));
   memcpy_P(EEPROM.getDataPtr(), &itemDefaults, sizeof(itemDefaults));
 }
 #endif
@@ -313,7 +313,7 @@ int storageSetup(void) {
         EEPROM.begin(sizeof(sto_data_t));
     #elif defined(ESP32)
         if (!EEPROM.begin(sizeof(sto_data_t))) {
-            Serial.printf("%s(): EEPROM initialization failed!\n", __FUNCTION__);
+            debugPrintf("%s(): EEPROM initialization failed!\n", __FUNCTION__);
             return -1;
         }
     #else
@@ -342,21 +342,21 @@ static inline int getNumber(sto_item_id_t itemId, T& itemValue) {
     int32_t itemAddr = getItemAddr(itemId, &maxItemSize);
 
     if (itemAddr < 0) {
-        Serial.printf("%s(): invalid item address!\n", __FUNCTION__);
+        debugPrintf("%s(): invalid item address!\n", __FUNCTION__);
         return -1;
     }
 
-    Serial.printf("%s(): addr=%i size=%u/%u\n", __FUNCTION__, itemAddr, sizeof(itemValue), maxItemSize);
+    debugPrintf("%s(): addr=%i size=%u/%u\n", __FUNCTION__, itemAddr, sizeof(itemValue), maxItemSize);
 
     if (sizeof(itemValue) != maxItemSize) {
-        Serial.printf("%s(): invalid item size (wrong data type)!\n", __FUNCTION__);
+        debugPrintf("%s(): invalid item size (wrong data type)!\n", __FUNCTION__);
         return -2;
     }
 
     EEPROM.get(itemAddr, itemValue);
 
     if (isEmpty(&itemValue, sizeof(itemValue))) { // item storage empty?
-        Serial.printf("%s(): storage empty -> returning default\n", __FUNCTION__);
+        debugPrintf("%s(): storage empty -> returning default\n", __FUNCTION__);
 
         memcpy_P(&itemValue, (PGM_P)&itemDefaults + itemAddr, sizeof(itemValue));  // set default value
     }
@@ -382,19 +382,19 @@ static inline int setNumber(sto_item_id_t itemId, const T& itemValue, bool commi
     int32_t itemAddr = getItemAddr(itemId, &maxItemSize);
 
     if (itemAddr < 0) {
-        Serial.printf("%s(): invalid item address!\n", __FUNCTION__);
+        debugPrintf("%s(): invalid item address!\n", __FUNCTION__);
         return -1;
     }
 
-    Serial.printf("%s(): addr=%i size=%u/%u commit=%u\n", __FUNCTION__, itemAddr, sizeof(itemValue), maxItemSize, commit);
+    debugPrintf("%s(): addr=%i size=%u/%u commit=%u\n", __FUNCTION__, itemAddr, sizeof(itemValue), maxItemSize, commit);
 
     if (sizeof(itemValue) != maxItemSize) {
-        Serial.printf("%s(): invalid item size (wrong data type)!\n", __FUNCTION__);
+        debugPrintf("%s(): invalid item size (wrong data type)!\n", __FUNCTION__);
         return -2;
     }
 
     if (isEmpty(&itemValue, sizeof(T))) {
-        Serial.printf("%s(): invalid item value!\n", __FUNCTION__);
+        debugPrintf("%s(): invalid item value!\n", __FUNCTION__);
         return -3;
     }
 
@@ -419,7 +419,7 @@ int storageGet(sto_item_id_t itemId, float& itemValue) {
 
     if (retCode != 0) return retCode;
 
-    Serial.printf("%s(): item=%i value=%.1f\n", __FUNCTION__, itemId, itemValue);
+    debugPrintf("%s(): item=%i value=%.1f\n", __FUNCTION__, itemId, itemValue);
 
     return 0;
 }
@@ -429,7 +429,7 @@ int storageGet(sto_item_id_t itemId, double& itemValue) {
 
     if (retCode != 0) return retCode;
 
-    Serial.printf("%s(): item=%i value=%.1f\n", __FUNCTION__, itemId, itemValue);
+    debugPrintf("%s(): item=%i value=%.1f\n", __FUNCTION__, itemId, itemValue);
 
     return 0;
 }
@@ -439,7 +439,7 @@ int storageGet(sto_item_id_t itemId, int8_t& itemValue) {
 
     if (retCode != 0) return retCode;
 
-    Serial.printf("%s(): item=%i value=%i\n", __FUNCTION__, itemId, itemValue);
+    debugPrintf("%s(): item=%i value=%i\n", __FUNCTION__, itemId, itemValue);
 
     return 0;
 }
@@ -449,7 +449,7 @@ int storageGet(sto_item_id_t itemId, int16_t& itemValue) {
 
     if (retCode != 0) return retCode;
 
-    Serial.printf("%s(): item=%i value=%i\n", __FUNCTION__, itemId, itemValue);
+    debugPrintf("%s(): item=%i value=%i\n", __FUNCTION__, itemId, itemValue);
 
     return 0;
 }
@@ -459,7 +459,7 @@ int storageGet(sto_item_id_t itemId, int32_t& itemValue) {
 
     if (retCode != 0) return retCode;
 
-    Serial.printf("%s(): item=%i value=%i\n", __FUNCTION__, itemId, itemValue);
+    debugPrintf("%s(): item=%i value=%i\n", __FUNCTION__, itemId, itemValue);
 
     return 0;
 }
@@ -469,7 +469,7 @@ int storageGet(sto_item_id_t itemId, uint8_t& itemValue) {
 
     if (retCode != 0) return retCode;
 
-    Serial.printf("%s(): item=%i value=%u\n", __FUNCTION__, itemId, itemValue);
+    debugPrintf("%s(): item=%i value=%u\n", __FUNCTION__, itemId, itemValue);
 
     return 0;
 }
@@ -477,7 +477,7 @@ int storageGet(sto_item_id_t itemId, uint8_t& itemValue) {
 int storageGet(sto_item_id_t itemId, uint16_t& itemValue) {
     int retCode = getNumber(itemId, itemValue);
     if (retCode != 0) return retCode;
-    Serial.printf("%s(): item=%i value=%u\n", __FUNCTION__, itemId, itemValue);
+    debugPrintf("%s(): item=%i value=%u\n", __FUNCTION__, itemId, itemValue);
 
     return 0;
 }
@@ -487,7 +487,7 @@ int storageGet(sto_item_id_t itemId, uint32_t& itemValue) {
 
     if (retCode != 0) return retCode;
 
-    Serial.printf("%s(): item=%i value=%u\n", __FUNCTION__, itemId, itemValue);
+    debugPrintf("%s(): item=%i value=%u\n", __FUNCTION__, itemId, itemValue);
 
     return 0;
 }
@@ -510,11 +510,11 @@ int storageGet(sto_item_id_t itemId, const char** itemValue) {
     itemSize = strlen(*itemValue) + 1;
 
     if (isEmpty(*itemValue, itemSize)) { // item storage empty?
-        Serial.printf("%s(): storage empty -> returning default\n", __FUNCTION__);
+        debugPrintf("%s(): storage empty -> returning default\n", __FUNCTION__);
         memcpy_P(&itemValue, (const void*)(&itemDefaults+itemAddr), itemSize);      // set default value
     }
 
-    Serial.printf("%s(): addr=%i size=%u item=%i value=\"%s\"\n", __FUNCTION__, itemAddr, itemSize, itemId, *itemValue);
+    debugPrintf("%s(): addr=%i size=%u item=%i value=\"%s\"\n", __FUNCTION__, itemAddr, itemSize, itemId, *itemValue);
 
     return 0;
 }
@@ -525,7 +525,7 @@ int storageGet(sto_item_id_t itemId, String& itemValue) {
     int32_t itemAddr = getItemAddr(itemId, &maxItemSize);
 
     if (itemAddr < 0) {
-        Serial.printf("%s(): invalid item address!\n", __FUNCTION__);
+        debugPrintf("%s(): invalid item address!\n", __FUNCTION__);
         return -1;
     }
 
@@ -536,7 +536,7 @@ int storageGet(sto_item_id_t itemId, String& itemValue) {
         if (isString(storageDataPtr, maxItemSize))  { // exist a null terminator?
             itemValue = String((const char*)storageDataPtr); // convert to C++ string
         } else {
-            Serial.printf("%s(): storage empty -> returning default\n", __FUNCTION__);
+            debugPrintf("%s(): storage empty -> returning default\n", __FUNCTION__);
             itemValue = String((PGM_P)&itemDefaults + itemAddr);  // set default string
         }
     #elif defined(ESP32)
@@ -547,14 +547,14 @@ int storageGet(sto_item_id_t itemId, String& itemValue) {
         if (isString(buf, maxItemSize)) { // exist a null terminator?
             itemValue = String((const char*)buf);
         } else {
-            Serial.printf("%s(): storage empty -> returning default\n", __FUNCTION__);
+            debugPrintf("%s(): storage empty -> returning default\n", __FUNCTION__);
             itemValue = String((PGM_P)&itemDefaults + itemAddr);  // set default string
         }
     #else
         #error("MCU not supported");
     #endif
 
-    Serial.printf("%s(): addr=%i size=%u item=%i value=\"%s\"\n", __FUNCTION__, itemAddr, itemValue.length() + 1, itemId, itemValue.c_str());
+    debugPrintf("%s(): addr=%i size=%u item=%i value=\"%s\"\n", __FUNCTION__, itemAddr, itemValue.length() + 1, itemId, itemValue.c_str());
 
     return 0;
 }
@@ -572,42 +572,42 @@ int storageGet(sto_item_id_t itemId, String& itemValue) {
  *         <0 - failed
  */
 int storageSet(sto_item_id_t itemId, float itemValue, bool commit) {
-    Serial.printf("%s(): item=%i value=%.1f\n", __FUNCTION__, itemId, itemValue);
+    debugPrintf("%s(): item=%i value=%.1f\n", __FUNCTION__, itemId, itemValue);
     return setNumber(itemId, itemValue, commit);
 }
 
 int storageSet(sto_item_id_t itemId, double itemValue, bool commit) {
-    Serial.printf("%s(): item=%i value=%.1f\n", __FUNCTION__, itemId, itemValue);
+    debugPrintf("%s(): item=%i value=%.1f\n", __FUNCTION__, itemId, itemValue);
     return setNumber(itemId, itemValue, commit);
 }
 
 int storageSet(sto_item_id_t itemId, int8_t itemValue, bool commit) {
-    Serial.printf("%s(): item=%i value=%i\n", __FUNCTION__, itemId, itemValue);
+    debugPrintf("%s(): item=%i value=%i\n", __FUNCTION__, itemId, itemValue);
     return setNumber(itemId, itemValue, commit);
 }
 
 int storageSet(sto_item_id_t itemId, int16_t itemValue, bool commit) {
-    Serial.printf("%s(): item=%i value=%i\n", __FUNCTION__, itemId, itemValue);
+    debugPrintf("%s(): item=%i value=%i\n", __FUNCTION__, itemId, itemValue);
     return setNumber(itemId, itemValue, commit);
 }
 
 int storageSet(sto_item_id_t itemId, int32_t itemValue, bool commit) {
-    Serial.printf("%s(): item=%i value=%i\n", __FUNCTION__, itemId, itemValue);
+    debugPrintf("%s(): item=%i value=%i\n", __FUNCTION__, itemId, itemValue);
     return setNumber(itemId, itemValue, commit);
 }
 
 int storageSet(sto_item_id_t itemId, uint8_t itemValue, bool commit) {
-    Serial.printf("%s(): item=%i value=%u\n", __FUNCTION__, itemId, itemValue);
+    debugPrintf("%s(): item=%i value=%u\n", __FUNCTION__, itemId, itemValue);
     return setNumber(itemId, itemValue, commit);
 }
 
 int storageSet(sto_item_id_t itemId, uint16_t itemValue, bool commit) {
-    Serial.printf("%s(): item=%i value=%u\n", __FUNCTION__, itemId, itemValue);
+    debugPrintf("%s(): item=%i value=%u\n", __FUNCTION__, itemId, itemValue);
     return setNumber(itemId, itemValue, commit);
 }
 
 int storageSet(sto_item_id_t itemId, uint32_t itemValue, bool commit) {
-    Serial.printf("%s(): item=%i value=%u\n", __FUNCTION__, itemId, itemValue);
+    debugPrintf("%s(): item=%i value=%u\n", __FUNCTION__, itemId, itemValue);
     return setNumber(itemId, itemValue, commit);
 }
 
@@ -617,15 +617,15 @@ int storageSet(sto_item_id_t itemId, const char* itemValue, bool commit) {
     int32_t itemAddr = getItemAddr(itemId, &maxItemSize);
 
     if (itemAddr < 0) {
-        Serial.printf("%s(): invalid item address!\n", __FUNCTION__);
+        debugPrintf("%s(): invalid item address!\n", __FUNCTION__);
         return -1;
     }
 
     valueSize = strlen(itemValue) + 1;
-    Serial.printf("%s(): item=%i value=\"%s\" addr=%i size=%u/%u\n", __FUNCTION__, itemId, itemValue, itemAddr, valueSize, maxItemSize);
+    debugPrintf("%s(): item=%i value=\"%s\" addr=%i size=%u/%u\n", __FUNCTION__, itemId, itemValue, itemAddr, valueSize, maxItemSize);
 
     if (valueSize > maxItemSize) { // invalid value size?
-        Serial.printf("%s(): string too large!\n", __FUNCTION__);
+        debugPrintf("%s(): string too large!\n", __FUNCTION__);
         return -2;
     }
 
@@ -650,7 +650,7 @@ int storageCommit(void) {
     int returnCode;
     // bool isTimerEnabled;
 
-    Serial.printf("%s(): save all data to NV memory\n", __FUNCTION__);
+    debugPrintf("%s(): save all data to NV memory\n", __FUNCTION__);
 
     /* While Flash memory erase/write operations no other code must be executed from
      * Flash! Since all code of the Timer1-ISR is placed in RAM, the interrupt does
@@ -681,7 +681,7 @@ int storageCommit(void) {
  *         <0 - failed
  */
 int storageFactoryReset(void) {
-    Serial.printf("%s(): reset all values\n", __FUNCTION__);
+    debugPrintf("%s(): reset all values\n", __FUNCTION__);
     memset(EEPROM.getDataPtr(), 0xFF, sizeof(sto_data_t));
 
     return storageCommit();
